@nielsdos 2025-05-07
MINIT is only executed once at startup of PHP, which is why the difference is so small. So this likely only has a real benefit on CLI, if at all. Not against this though as this looks a bit simpler, but I'll let bukka judge.

> don't have access to a big-endian system to test on. hopefully Github CI does.

Nightly has a BE job.

---

I was also looking into optimizing pack() after I saw your docs PR. I noted that getting rid of convert_to_long provided a nice benefit and I have a local patch for that. I'll submit that later today.