@nielsdos 2025-05-07
I believe it's possible to get rid of the mapping arrays altogether and just use some byte swapping, didn't test on a big endian system yet however: https://gist.github.com/nielsdos/db426e7776a7e73fc0b953c082d91e76