@divinity76 2025-05-07
btw I wonder at what point big-endian becomes so obscure that support can be dropped entirely,

does anyone actually run PHP on a big-endian system?