@nielsdos 2025-05-07
> btw I wonder at what point big-endian becomes so obscure that support can be dropped entirely,

Some protocols may use big endian in which case you'll still want to have this support.

> does anyone actually run PHP on a big-endian system?

Apparently yes, not too long ago we fixed some issues with that+FFI.