@divinity76 2025-05-08
>Some protocols may use big endian in which case you'll still want to have this support.

I didn't mean "remove little-endian-php's ability to parse big-endian bytes", I meant more like "remove php-src's compile-on-big-endian compatibility stuff"

>Apparently yes, not too long ago we fixed some issues with that+FFI.

Interesting, do you remember the title/issue/link?