@Girgias 2025-05-08
> btw I wonder at what point big-endian becomes so obscure that support can be dropped entirely,
> 
> does anyone actually run PHP on a big-endian system?

People running PHP on IBM Z mainframes, we used to have an S390X CI pipeline on Travis before that company made it impossible for us to use them. And some IBM folks do still submit fixes when they encounter bugs.