@nielsdos 2025-05-08
> > Some protocols may use big endian in which case you'll still want to have this support.
> 
> I didn't mean "remove little-endian-php's ability to parse big-endian bytes", I meant more like "remove php-src's compile-on-big-endian compatibility stuff"
> 

Ah right. Often, there isn't really a lot of extra code necessary in PHP to support big endian machines, so it's not too big of a burden. We have some here, in bcmath and in phar, but they're not a lot.

> > Apparently yes, not too long ago we fixed some issues with that+FFI.
> 
> Interesting, do you remember the title/issue/link?

This was one of them: https://github.com/php/php-src/issues/16013