@nielsdos 2025-06-10
I had a "competing" PR at https://github.com/php/php-src/pull/18524 that did away with the map entirely.
Initially it showed an improvement in performance but now it does not (i.e. it's sometimes slower) and I haven't been able to figure out why. If I can't get that to improve I think we should go with your approach.