@divinity76 2025-06-10
how much slower? got some numbers?