@divinity76 2025-06-10
fwiw, unless ZEND_BYTES_SWAP32 is sluggish, i see no reason for your PR https://github.com/php/php-src/pull/18524 to be slow