@nielsdos 2025-06-10
I'm going to do a fresh recompile on a rebase on master. Maybe all that incremental building and swapping back&forth between branches messed up something.